move ui_actions service to package
