+++ packages/kbn-ui-actions-browser/src/actions/index.ts
@@ -6,11 +6,6 @@
  * Side Public License, v 1.
  */
 
-import { Trigger } from '@kbn/ui-actions-browser';
-
-export const VISUALIZE_FIELD_TRIGGER = 'VISUALIZE_FIELD_TRIGGER';
-export const visualizeFieldTrigger: Trigger = {
-  id: VISUALIZE_FIELD_TRIGGER,
-  title: 'Visualize field',
-  description: 'Triggered when user wants to visualize a field.',
-};
+export * from './action';
+export * from './action_internal';
+export * from './incompatible_action_error';
