+++ packages/kbn-ui-actions-browser/src/context_menu/build_eui_context_menu_panels.tsx
