+++ packages/kbn-ui-actions-browser/src/context_menu/open_context_menu.test.ts
