+++ packages/kbn-ui-actions-browser/src/index.ts
@@ -7,3 +7,6 @@
  */
 
 export * from './types';
+export * from './triggers';
+export * from './actions';
+export * from './service';
