+++ packages/kbn-ui-actions-browser/src/triggers/trigger_contract.ts
