+++ packages/kbn-ui-actions-browser/src/types/index.ts
@@ -7,3 +7,4 @@
  */
 
 export * from './presentable';
+export * from './registries';
