+++ src/plugins/ui_actions/public/actions/index.ts
@@ -6,7 +6,4 @@
  * Side Public License, v 1.
  */
 
-export * from './action';
-export * from './action_internal';
 export * from './create_action';
-export * from './incompatible_action_error';
