+++ src/plugins/ui_actions/public/tests/get_trigger_compatible_actions.test.ts
@@ -8,8 +8,7 @@
 
 import { uiActionsPluginMock } from '../mocks';
 import { createHelloWorldAction } from './test_samples';
-import { ActionDefinition } from '../actions';
-import { Trigger } from '../triggers';
+import { Trigger, ActionDefinition } from '@kbn/ui-actions-browser';
 import { OverlayStart } from '@kbn/core/public';
 
 let action: ActionDefinition<{ name: string }>;
