+++ src/plugins/ui_actions/public/triggers/index.ts
@@ -6,10 +6,6 @@
  * Side Public License, v 1.
  */
 
-export * from './trigger';
-export * from './trigger_contract';
-export * from './trigger_internal';
 export * from './row_click_trigger';
 export * from './visualize_field_trigger';
 export * from './visualize_geo_field_trigger';
-export * from './default_trigger';
