+++ src/plugins/ui_actions/public/triggers/row_click_trigger.ts
@@ -8,7 +8,7 @@
 
 import { i18n } from '@kbn/i18n';
 import type { Datatable } from '@kbn/expressions-plugin/common';
-import { Trigger } from '.';
+import { Trigger } from '@kbn/ui-actions-browser';
 
 export const ROW_CLICK_TRIGGER = 'ROW_CLICK_TRIGGER';
