+++ src/plugins/ui_actions/public/types.ts
@@ -8,8 +8,7 @@
 
 import type { AggregateQuery } from '@kbn/es-query';
 import type { DataViewSpec } from '@kbn/data-views-plugin/public';
-import { ActionInternal } from './actions/action_internal';
-import { TriggerInternal } from './triggers/trigger_internal';
+import { ActionInternal, TriggerInternal } from '@kbn/ui-actions-browser';
 
 export type TriggerRegistry = Map<string, TriggerInternal<object>>;
 export type ActionRegistry = Map<string, ActionInternal>;
